refactor: サービス共通化
