+++ backend/src/application/query/dto/getUserPost.dto.ts
@@ -6,7 +6,6 @@ export class GetUserPostRequestDto {
   }
 }
 
-
 export type PostCommentDto = {
   readonly id: string
   readonly userId: string
