+++ backend/src/application/query/service/getUserPost.service.ts
@@ -2,8 +2,7 @@ import { Post } from "../../../domain/post/entity/post.entity";
 import { User } from "../../../domain/user/entity/user.entity";
 import { PostQueryService } from "../../post/service/postQuery.service";
 import { UserQueryService } from "../../user/service/userQuery.service";
-import { getAllPostWithUserDto, PostCommentDto } from "../dto/getAllPostWithUser.dto";
-import { GetUserPostRequestDto, GetUserPostResponseDto } from "../dto/getUserPost.dto";
+import { GetUserPostRequestDto, GetUserPostResponseDto, PostCommentDto } from "../dto/getUserPost.dto";
 
 export class GetUserPostService {
   constructor(
@@ -11,7 +10,7 @@ export class GetUserPostService {
     private readonly userQueryService: UserQueryService
   ) { }
 
-  public async execute(getUserPostRequestDto: GetUserPostRequestDto): Promise<getAllPostWithUserDto[]> {
+  public async getAllPostByUser(getUserPostRequestDto: GetUserPostRequestDto): Promise<GetUserPostResponseDto[]> {
     const userId: string = getUserPostRequestDto.userId;
     const posts: Post[] = await this.postQueryService.findAllByUserId(userId);
     const users: User[] = await this.userQueryService.findAll();
@@ -36,6 +35,31 @@ export class GetUserPostService {
       ));
     }
 
+    return result;
+  }
+
+  public async getAllPost(): Promise<GetUserPostResponseDto[]> {
+    const posts: Post[] = await this.postQueryService.findAll();
+    const users: User[] = await this.userQueryService.findAll();
+
+    const result: GetUserPostResponseDto[] = [];
+    for (const post of posts) {
+      const user = users.find((user) => user.getUserId().value === post.getUserId().value);
+      if (!user) {
+        throw new Error("ユーザーが見つかりませんでした");
+      }
+      result.push(new GetUserPostResponseDto(
+        post.getPostId().value,
+        post.getTitle().value,
+        post.getPrice().value,
+        post.getUserId().value,
+        user.getUsername().value,
+        post.getCreatedAt(),
+        post.getUpdatedAt(),
+        post.getLikes().map(like => like.getUserId().value),
+        this.getComments(post, users)
+      ));
+    }
     return result;
   }
